+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Support for additional functionals. (ORCA)
 - Support for coupled cluster calculations. (ORCA)
 - Track Hartree&ndash;Fock type for ab initio methods. (ORCA)
+- A script to combine JSON files recursively.
 
 ### Changed